move decode_test to tests folder
