+++ .github/workflows/test.yml
@@ -53,7 +53,7 @@ jobs:
         run: curl -sL firebase.tools | bash
       
       - name: Run Go tests with Firebase emulator
-        run: firebase emulators:exec --only firestore 'cd ./apps/question-service/tests; go test ./...'
+        run: firebase emulators:exec --only firestore 'cd ./apps/question-service; go test -v ./tests'
 
   frontend-unit-tests:
     runs-on: ubuntu-latest
